[Refactor] SectionSettings struct

+++ GreenhouseApp/src/SectionWidget/settingsdialog.cpp
@@ -24,13 +24,13 @@ SectionSettings SettingsDialog::downloadParameters() const
 {
     SectionSettings sectionSettings;
     sectionSettings.section_name = ui->le_sectionName->text();
-    sectionSettings.sectionID = ui->sb_id->value();
+    sectionSettings.sectionID = static_cast<quint8>(ui->sb_id->value());
     sectionSettings.picture_path = ui->lbl_fotoPath->text();
-    sectionSettings.time_hours = ui->sb_timeHours->value();
-    sectionSettings.time_minutes = ui->sb_timeMinutes->value();
-    sectionSettings.temperature_expected = ui->sb_maxTemp->value();
-    sectionSettings.humidity_expected = ui->sb_minHum->value();
-    sectionSettings.insolation_expected = ui->sb_minIns->value();
+    sectionSettings.time_hours = static_cast<quint8>(ui->sb_timeHours->value());
+    sectionSettings.time_minutes = static_cast<quint8>(ui->sb_timeMinutes->value());
+    sectionSettings.temperature_expected = static_cast<quint8>(ui->sb_maxTemp->value());
+    sectionSettings.humidity_expected = static_cast<quint8>(ui->sb_minHum->value());
+    sectionSettings.insolation_expected = static_cast<quint8>(ui->sb_minIns->value());
     return sectionSettings;
 }
