+++ GreenhouseApp/src/SectionWidget/settingsdialog.h
@@ -12,11 +12,11 @@ struct SectionSettings{
     quint8 sectionID = 0;
     QString section_name = "Section";
     QString picture_path;
-    int temperature_expected = 0;
-    int humidity_expected = 0;
-    int insolation_expected = 0;
-    int time_hours = 0;
-    int time_minutes = 0;
+    quint8 temperature_expected = 0;
+    quint8 humidity_expected = 0;
+    quint8 insolation_expected = 0;
+    quint8 time_hours = 0;
+    quint8 time_minutes = 0;
 };
 
 class SettingsDialog : public QDialog
